+++ src/infrastructure/applications/console/config.ts
@@ -1,11 +1,11 @@
-import {PostgresConnectionOptions} from '@steroidsjs/typeorm/driver/postgres/PostgresConnectionOptions';
+import {PostgresDataSourceOptions} from 'typeorm/driver/postgres/PostgresDataSourceOptions';
 import * as path from 'path';
 import * as fs from 'node:fs';
+import {CommandModule} from 'nestjs-command';
 import baseConfig from '../base/config';
 import {IConsoleAppModuleConfig} from './IConsoleAppModuleConfig';
 import {EntityCodeGenerateCommand} from '../../commands/entity-generator/EntityCodeGenerateCommand';
 import {MigrateCommand} from '../../commands/MigrateCommand';
-import {CommandModule} from 'nestjs-command';
 
 const isMigrateCommand = !!(process.argv || []).find(arg => /^migrate/.exec(arg));
 
@@ -28,7 +28,7 @@ export default {
         const config = baseConfig.config();
 
         // For deployment to use files in dist directory.
-        const envRootDir = process.env.APP_ENVIRONMENT === 'dev' ? 'src' : 'dist';        
+        const envRootDir = process.env.APP_ENVIRONMENT === 'dev' ? 'src' : 'dist';
 
         // If CLI_PATH is specified then use directory from it.
         const migrationsRootDir = process.env.CLI_PATH && path.dirname(process.env.CLI_PATH)
@@ -43,7 +43,7 @@ export default {
                     ? collectMigrations(migrationsRootDir)
                     : [], // Do not include migrations on web and other cli commands
                 migrationsTableName: 'migrations',
-            } as PostgresConnectionOptions
+            } as PostgresDataSourceOptions
         } as IConsoleAppModuleConfig;
     },
     module: (config: IConsoleAppModuleConfig) => {